ATR-662: refactoring - moved PX1042 diagnostic walker to a separate file
